refactor: split up hardware checks, added memory c
