+++ files/usr/etc/ublue-update/ublue-update.toml
@@ -1,5 +1,9 @@
 [checks]
     min_battery_percent = 50.0
-    max_cpu_load = 50.0
+    max_cpu_load_percent = 50.0
+    max_mem_percent = 50.0
 [notify]
     dbus_notify = true
+[schedule]
+    frequency="weekly"
+
