+++ src/ublue_update/cli.py
@@ -1,25 +1,25 @@
-import psutil
 import os
 import subprocess
 import logging
-import tomllib
 import argparse
-
 
 from ublue_update.update_checks.system import system_update_check
 from ublue_update.update_checks.wait import transaction_wait
+from ublue_update.update_inhibitors.hardware import check_hardware_inhibitors
+from ublue_update.config import load_value
 
 
-def notify(title: str, body: str, actions: list = [], expire_time: int = 0):
+def notify(title: str, body: str, actions: list = [], urgency: str = "normal"):
+    if not dbus_notify:
+        return
     args = [
         "/usr/bin/notify-send",
         title,
         body,
         "--app-name=Universal Blue Updater",
         "--icon=software-update-available-symbolic",
+        f"--urgency=${urgency}",
     ]
-    if expire_time != 0:
-        args.append(f"--expire-time={expire_time}")
     if actions != []:
         for action in actions:
             args.append(f"--action={action}")
@@ -28,11 +28,13 @@ def notify(title: str, body: str, actions: list = [], expire_time: int = 0):
 
 
 def ask_for_updates():
+    if not dbus_notify:
+        return
     out = notify(
         "System Updater",
         "Update available, but system checks failed. Update now?",
         ["universal-blue-update-confirm=Confirm"],
-        15000,
+        "critical"
     )
     # if the user has confirmed
     if "universal-blue-update-confirm" in out.stdout.decode("utf-8"):
@@ -51,103 +53,17 @@ def check_for_updates(checks_failed: bool) -> bool:
     return False
 
 
-def check_cpu_load() -> dict:
-    # get load average percentage in last 5 minutes:
-    # https://psutil.readthedocs.io/en/latest/index.html?highlight=getloadavg
-    cpu_load = psutil.getloadavg()[1] / psutil.cpu_count() * 100
-    return {
-        "passed": cpu_load < max_cpu_load,
-        "message": f"CPU load is above {max_cpu_load}%",
-    }
-
-
-def check_network_status() -> dict:
-    network_status = psutil.net_if_stats()
-    # check each network interface
-    network_up = False
-    for key in network_status.keys():
-        if key != "lo":
-            if network_status[key][0]:
-                network_up = True
-                break
-    return {"passed": network_up, "message": "Network not enabled"}
-
-
-def check_battery_status() -> dict:
-    battery_status = psutil.sensors_battery()
-    # null safety on the battery variable, it returns "None"
-    # when the system doesn't have a battery
-    battery_pass: bool = True
-    if battery_status is not None:
-        battery_pass = (
-            battery_status.percent > min_battery_percent or battery_status.power_plugged
-        )
-    return {
-        "passed": battery_pass,
-        "message": f"Battery less than {min_battery_percent}%",
-    }
-
-
 def hardware_inhibitor_checks_failed(
-    hardware_checks_failed: bool, failures: list, dbus_ask_for_updates: bool
+    hardware_checks_failed: bool, failures: list, hardware_check: bool
 ):
     # ask if an update can be performed through dbus notifications
-    if check_for_updates(hardware_checks_failed) and dbus_ask_for_updates:
+    if check_for_updates(hardware_checks_failed) and not hardware_check:
         log.info("Harware checks failed, but update is available")
         ask_for_updates()
     # notify systemd that the checks have failed,
     # systemd will try to rerun the unit
     exception_log = "\n - ".join(failures)
     raise Exception(f"update failed to pass checks: \n - {exception_log}")
-
-
-def check_hardware_inhibitors() -> bool:
-
-    hardware_inhibitors = [
-        check_network_status(),
-        check_battery_status(),
-        check_cpu_load(),
-    ]
-
-    failures = []
-    hardware_checks_failed = False
-    for inhibitor_result in hardware_inhibitors:
-        if not inhibitor_result["passed"]:
-            hardware_checks_failed = True
-            failures.append(inhibitor_result["message"])
-    if not hardware_checks_failed:
-        log.info("System passed hardware checks")
-    return hardware_checks_failed, failures
-
-
-def load_config():
-    # load config values
-    config_paths = [
-        os.path.expanduser("~/.config/ublue-update/ublue-update.toml"),
-        "/etc/ublue-update/ublue-update.toml",
-        "/usr/etc/ublue-update/ublue-update.toml",
-    ]
-
-    # search for the right config
-    config_path = ""
-    fallback_config_path = ""
-    for path in config_paths:
-        if os.path.isfile(path):
-            if config_path == "":
-                config_path = path
-            fallback_config_path = path
-            break
-
-    fallback_config = tomllib.load(open(fallback_config_path, "rb"))
-    config = tomllib.load(open(config_path, "rb"))
-    return config, fallback_config
-
-
-def load_value(key, value):
-    fallback = fallback_config[key][value]
-    if key in config.keys():
-        return config[key].get(value, fallback)
-    return fallback
 
 
 def run_updates():
@@ -164,6 +80,7 @@ def run_updates():
 
             executable = os.access(full_path, os.X_OK)
             if executable:
+                log.info(f"Running update script: {full_path}")
                 out = subprocess.run(
                     [full_path], stdout=subprocess.PIPE, stderr=subprocess.STDOUT
                 )
@@ -171,27 +88,21 @@ def run_updates():
                 if out.returncode != 0:
                     log.info(f"{full_path} returned error code: {out.returncode}")
                     log.info(f"Program output: \n {out.stdout}")
-                    if dbus_notify:
-                        notify(
-                            "System Updater",
-                            f"Error in update script: {file}, check logs for more info",
-                        )
+                    notify(
+                        "System Updater",
+                        f"Error in update script: {file}, check logs for more info",
+                    )
             else:
                 log.info(f"could not execute file {full_path}")
-    if dbus_notify:
-        notify(
-            "System Updater",
-            "System update complete, reboot for changes to take effect",
-        )
+    notify(
+        "System Updater",
+        "System update complete, reboot for changes to take effect",
+    )
     log.info("System update complete")
     os._exit(0)
 
 
-config, fallback_config = load_config()
-
 dbus_notify: bool = load_value("notify", "dbus_notify")
-min_battery_percent: int = load_value("checks", "min_battery_percent")
-max_cpu_load: int = load_value("checks", "max_cpu_load")
 
 # setup logging
 logging.basicConfig(
@@ -199,6 +110,7 @@ logging.basicConfig(
     level=os.getenv("UBLUE_LOG", default=logging.INFO),
 )
 log = logging.getLogger(__name__)
+
 
 def main():
 
@@ -238,7 +150,7 @@ def main():
             hardware_inhibitor_checks_failed(
                 hardware_checks_failed,
                 failures,
-                dbus_notify and not args.check,
+                args.check,
             )
         if args.check:
             os._exit(0)
@@ -251,9 +163,8 @@ def main():
 
     # system checks passed
     log.info("System passed all update checks")
-    if dbus_notify:
-        notify(
-            "System Updater",
-            "System passed checks, updating ...",
-        )
+    notify(
+        "System Updater",
+        "System passed checks, updating ...",
+    )
     run_updates()
