+++ src/ublue_update/update_checks/wait.py
@@ -5,7 +5,7 @@ from time import sleep
 
 def transaction():
     """Pull deployment status via rpm-ostree"""
-    rpm_ostree_status = ["rpm-ostree", "status",  "--json"]
+    rpm_ostree_status = ["rpm-ostree", "status", "--json"]
     status = run(rpm_ostree_status, stdout=PIPE)
     """Parse transaction state"""
     return loads(status.stdout)["transaction"]
